refactor: move api key utilities to seperate file
